refactor(archival): one home for the archiefactiedatum calculation
phpmd failed #3592 on RetentionService: 1178 lines of code against a threshold
of 1000, and calculateArchiveActionDate() at the cyclomatic ceiling. Both were
mine; the branch point was clean.

Fixing that turned up the better reason to move the code. There were TWO
implementations of the archiefactiedatum question and only one of them ran.
`ArchiveActionDateCalculator` carried its own `determineBrondatum` and
`brondatumFromProperty`, plus a unit test suite that made it look live, and had
ZERO production callers anywhere in the repository: every real calculation
went through RetentionService's private copy. A duplicate with its own green
tests is worse than no duplicate, because the next person to fix a derivation
bug has even odds of fixing the copy nobody calls.

So the live implementation moves INTO that class, under that name, and the
dead body and its stale test go. `ArchiveActionDateCalculator::calculate()` now
owns the whole question: which derivation method applies, where the retention
period starts counting from, what to do when that cannot be answered, and the
procestermijn for `termijn`. RetentionService keeps a delegating
`calculateArchiveActionDate()` because callers and the spec both name it.

On the way the code got simpler, not just shorter:

- the `eigenschap` and closure-field branches each carried their own
  copy of the parse-and-log block that `brondatumFromProperty` already had.
  They now share it through a quiet `brondatumFromOptionalProperty`, which
  returns silence for an unconfigured optional property because those methods
  have a creation-date fallback and always did;
- the refusal-or-fallback decision and the procestermijn step are named
  methods, so `calculate()` reads as the sequence it is.

The test harness builds the REAL calculator over the same mapper mock the
derivation tests program, not a mock of its own: those tests assert what the
calculator does with a relation, and a mocked calculator would assert only
that RetentionService called it.

The constructor now takes ten collaborators, one over phpmd's
ExcessiveParameterList threshold, and carries a documented suppression. The
tenth arrived by extracting logic out of this class rather than adding any, and
collapsing them behind a parameter object would hide the dependency graph
without reducing it.

Verified: phpunit 19564/19564 across the whole unit suite, phpcs exit 0, phpmd
both legs exit 0 with zero findings, phpstan exit 0 no errors, psalm exit 0 no
errors.
